[test, hmac] Merge the tests `hmac_enc` and `hmac_irq`
