+++ hw/top_earlgrey/dv/chip_sim_cfg.hjson
@@ -737,12 +737,6 @@
       run_opts: ["+sw_test_timeout_ns=15000000"]
     }
     {
-      name: chip_sw_hmac_enc_irq
-      uvm_test_seq: chip_sw_base_vseq
-      sw_images: ["sw/device/tests/hmac_enc_irq_test:1"]
-      en_run_modes: ["sw_test_mode_test_rom"]
-    }
-    {
       name: chip_sw_hmac_enc
       uvm_test_seq: chip_sw_base_vseq
       sw_images: ["sw/device/tests/hmac_enc_test:1"]
